+++ app/src/main/java/woowacourse/shopping/module/DateFormatterModule.kt
@@ -5,8 +5,8 @@ import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.android.components.ActivityComponent
 import dagger.hilt.android.scopes.ActivityScoped
-import woowacourse.shopping.ui.cart.DateFormatter
-import woowacourse.shopping.ui.cart.KoreanLocaleDateFormatter
+import woowacourse.shopping.presentation.cart.DateFormatter
+import woowacourse.shopping.presentation.cart.KoreanLocaleDateFormatter
 
 @Module
 @InstallIn(ActivityComponent::class)
@@ -14,6 +14,6 @@ abstract class DateFormatterModule {
     @Binds
     @ActivityScoped
     abstract fun dateFormatter(
-        dateFormatter: KoreanLocaleDateFormatter
-    ): DateFormatter
+        dateFormatter: woowacourse.shopping.presentation.cart.KoreanLocaleDateFormatter
+    ): woowacourse.shopping.presentation.cart.DateFormatter
 }
