--- app/src/main/java/woowacourse/shopping/ui/bindingadapter/ImageBinding.kt
@@ -1,12 +1,0 @@
-package woowacourse.shopping.ui.bindingadapter
-
-import android.widget.ImageView
-import androidx.databinding.BindingAdapter
-import com.bumptech.glide.Glide
-
-@BindingAdapter("imageUrl")
-fun loadImage(view: ImageView, url: String) {
-    Glide.with(view.context)
-        .load(url)
-        .into(view)
-}
