--- app/src/main/res/layout/fragment_cart.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools">
-
-    <data>
-
-        <variable
-            name="vm"
-            type="woowacourse.shopping.ui.cart.CartViewModel" />
-    </data>
-
-
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rv_cart_products"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
-        tools:listitem="@layout/item_cart_product" />
-
-
-</layout>
