+++ presentation/src/main/java/woowacourse/shopping/presentation/MainActivity.kt
@@ -1,0 +1,75 @@
+package woowacourse.shopping.presentation
+
+import android.content.Intent
+import android.os.Bundle
+import android.view.Menu
+import android.widget.Toast
+import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.ViewModelProvider
+import dagger.hilt.android.AndroidEntryPoint
+import woowacourse.shopping.presentation.cart.CartActivity
+import woowacourse.shopping.presentation.databinding.ActivityMainBinding
+
+@AndroidEntryPoint
+class MainActivity : AppCompatActivity() {
+
+    private val binding by lazy { ActivityMainBinding.inflate(layoutInflater) }
+
+    private val viewModel by lazy {
+        ViewModelProvider(this)[MainViewModel::class.java]
+    }
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(binding.root)
+
+        setupBinding()
+        setupToolbar()
+        setupView()
+    }
+
+
+    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
+        menuInflater.inflate(R.menu.cart_menu, menu)
+        menu?.findItem(R.id.cart)?.actionView?.let { view ->
+            view.setOnClickListener { navigateToCart() }
+        }
+        return true
+    }
+
+    private fun setupBinding() {
+        binding.lifecycleOwner = this
+        binding.vm = viewModel
+    }
+
+    private fun setupToolbar() {
+        setSupportActionBar(binding.toolbar)
+    }
+
+    private fun setupView() {
+        setupProductData()
+        setupProductList()
+    }
+
+    private fun setupProductData() {
+        viewModel.getAllProducts()
+    }
+
+    private fun setupProductList() {
+        viewModel.products.observe(this) {
+            val adapter = ProductAdapter(
+                items = it,
+                onClickProduct = viewModel::addCartProduct
+            )
+            binding.rvProducts.adapter = adapter
+        }
+        viewModel.onProductAdded.observe(this) {
+            if (!it) return@observe
+            Toast.makeText(this, getString(R.string.cart_added), Toast.LENGTH_SHORT).show()
+        }
+    }
+
+    private fun navigateToCart() {
+        startActivity(Intent(this, CartActivity::class.java))
+    }
+}
