+++ presentation/src/main/java/woowacourse/shopping/presentation/MainViewModel.kt
@@ -1,0 +1,38 @@
+package woowacourse.shopping.presentation
+
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
+import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.launch
+import woowacourse.shopping.data.CartRepository
+import woowacourse.shopping.data.ProductRepository
+import woowacourse.shopping.data.model.Product
+import javax.inject.Inject
+
+@HiltViewModel
+class MainViewModel @Inject constructor(
+    private val productRepository: ProductRepository,
+    private val cartRepository: CartRepository
+) : ViewModel() {
+
+    private val _products: MutableLiveData<List<Product>> = MutableLiveData(emptyList())
+    val products: LiveData<List<Product>> get() = _products
+
+    private val _onProductAdded: MutableLiveData<Boolean> = MutableLiveData(false)
+    val onProductAdded: LiveData<Boolean> get() = _onProductAdded
+
+
+    fun addCartProduct(product: Product) {
+        viewModelScope.launch {
+            cartRepository.addCartProduct(product)
+            _onProductAdded.value = true
+        }
+    }
+
+    fun getAllProducts() {
+        _products.value = productRepository.getAllProducts()
+    }
+
+}
