+++ presentation/src/main/java/woowacourse/shopping/presentation/ProductAdapter.kt
@@ -1,0 +1,21 @@
+package woowacourse.shopping.presentation
+
+import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
+import woowacourse.shopping.data.model.Product
+
+class ProductAdapter(
+    private val items: List<Product>,
+    private val onClickProduct: (Product) -> Unit
+) : RecyclerView.Adapter<ProductViewHolder>() {
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ProductViewHolder {
+        return ProductViewHolder.from(parent, onClickProduct)
+    }
+
+    override fun onBindViewHolder(holder: ProductViewHolder, position: Int) {
+        holder.bind(items[position])
+    }
+
+    override fun getItemCount(): Int = items.size
+}
