+++ presentation/src/main/java/woowacourse/shopping/presentation/bindingadapter/ImageBinding.kt
@@ -1,0 +1,12 @@
+package woowacourse.shopping.presentation.bindingadapter
+
+import android.widget.ImageView
+import androidx.databinding.BindingAdapter
+import com.bumptech.glide.Glide
+
+@BindingAdapter("imageUrl")
+fun loadImage(view: ImageView, url: String) {
+    Glide.with(view.context)
+        .load(url)
+        .into(view)
+}
