+++ presentation/src/main/java/woowacourse/shopping/presentation/cart/CartProductAdapter.kt
@@ -1,0 +1,34 @@
+package woowacourse.shopping.presentation.cart
+
+import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
+import woowacourse.shopping.data.model.Product
+
+class CartProductAdapter(
+    items: List<Product>,
+    onClickDelete: (position: Int) -> Unit,
+    private val dateFormatter: DateFormatter,
+) : RecyclerView.Adapter<CartProductViewHolder>() {
+
+    private val items: MutableList<Product> = items.toMutableList()
+
+    private val onClickDelete = { position: Int ->
+        onClickDelete(position)
+        removeItem(position)
+    }
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): CartProductViewHolder {
+        return CartProductViewHolder.from(parent, dateFormatter, onClickDelete)
+    }
+
+    override fun onBindViewHolder(holder: CartProductViewHolder, position: Int) {
+        holder.bind(items[position])
+    }
+
+    override fun getItemCount(): Int = items.size
+
+    private fun removeItem(position: Int) {
+        items.removeAt(position)
+        notifyItemRemoved(position)
+    }
+}
