+++ presentation/src/main/java/woowacourse/shopping/presentation/cart/CartProductViewHolder.kt
@@ -1,0 +1,38 @@
+package woowacourse.shopping.presentation.cart
+
+import android.view.LayoutInflater
+import android.view.ViewGroup
+import androidx.recyclerview.widget.RecyclerView
+import woowacourse.shopping.data.model.Product
+import woowacourse.shopping.presentation.databinding.ItemCartProductBinding
+
+class CartProductViewHolder(
+    private val binding: ItemCartProductBinding,
+    private val dateFormatter: DateFormatter,
+    onClickDelete: (position: Int) -> Unit,
+) : RecyclerView.ViewHolder(binding.root) {
+
+    init {
+        binding.ivCartProductDelete.setOnClickListener {
+            val position = adapterPosition
+            onClickDelete(position)
+        }
+    }
+
+    fun bind(product: Product) {
+        binding.item = product
+        binding.tvCartProductCreatedAt.text = dateFormatter.formatDate(product.createdAt)
+    }
+
+    companion object {
+        fun from(
+            parent: ViewGroup,
+            dateFormatter: DateFormatter,
+            onClickDelete: (position: Int) -> Unit,
+        ): CartProductViewHolder {
+            val binding = ItemCartProductBinding
+                .inflate(LayoutInflater.from(parent.context), parent, false)
+            return CartProductViewHolder(binding, dateFormatter, onClickDelete)
+        }
+    }
+}
