+++ presentation/src/main/res/drawable/ic_cart.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="27dp"
+    android:height="24dp"
+    android:viewportWidth="27"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M2.003,19.409L0.004,9.846C-0.001,9.804 -0.001,9.761 0.004,9.719C0.004,9.578 0.056,9.443 0.15,9.343C0.244,9.243 0.371,9.188 0.504,9.188H16.531L17.241,6.377C17.272,6.268 17.336,6.172 17.422,6.104C17.508,6.036 17.613,6 17.721,6H21.5C21.633,6 21.76,6.056 21.854,6.156C21.947,6.255 22,6.39 22,6.531C22,6.672 21.947,6.807 21.854,6.907C21.76,7.007 21.633,7.063 21.5,7.063H18.101L15.001,19.435C14.98,19.509 14.944,19.578 14.896,19.635C14.848,19.693 14.788,19.739 14.721,19.77C14.908,20.104 15.005,20.486 15.001,20.875C15.001,21.439 14.79,21.979 14.415,22.378C14.04,22.776 13.532,23 13.002,23C12.471,23 11.962,22.776 11.587,22.378C11.212,21.979 11.002,21.439 11.002,20.875C11.004,20.501 11.099,20.135 11.277,19.813H5.728C5.906,20.135 6,20.501 6.003,20.875C6.003,21.439 5.792,21.979 5.417,22.378C5.042,22.776 4.533,23 4.003,23C3.473,23 2.964,22.776 2.589,22.378C2.214,21.979 2.003,21.439 2.003,20.875C2.005,20.483 2.109,20.098 2.303,19.765C2.231,19.734 2.166,19.686 2.113,19.625C2.061,19.563 2.024,19.489 2.003,19.409ZM4.058,10.25H1.128L1.573,12.375H4.278L4.058,10.25ZM16.266,10.25H12.951L12.731,12.375H15.731L16.266,10.25ZM14.926,15.563L15.461,13.438H12.616L12.392,15.563H14.926ZM14.656,16.625H12.282L12.062,18.75H14.116L14.656,16.625ZM13.002,21.938C13.199,21.938 13.392,21.875 13.557,21.758C13.721,21.642 13.849,21.476 13.925,21.282C14.001,21.087 14.021,20.874 13.982,20.668C13.943,20.462 13.848,20.272 13.708,20.124C13.569,19.975 13.391,19.874 13.196,19.833C13.003,19.792 12.802,19.813 12.619,19.893C12.436,19.974 12.28,20.11 12.17,20.285C12.06,20.459 12.002,20.665 12.002,20.875C12.002,21.157 12.107,21.427 12.295,21.626C12.482,21.826 12.736,21.938 13.002,21.938ZM9.002,18.75H11.057L11.277,16.625H9.002V18.75ZM9.002,15.563H11.387L11.612,13.438H9.002V15.563ZM9.002,12.375H11.722L11.942,10.25H9.002V12.375ZM5.948,18.75H8.002V16.625H5.728L5.948,18.75ZM5.613,15.563H8.002V13.438H5.393L5.613,15.563ZM5.278,12.375H8.002V10.25H5.063L5.278,12.375ZM4.003,21.938C4.201,21.938 4.394,21.875 4.558,21.758C4.723,21.642 4.851,21.476 4.927,21.282C5.002,21.087 5.022,20.874 4.984,20.668C4.945,20.462 4.85,20.272 4.71,20.124C4.57,19.975 4.392,19.874 4.198,19.833C4.004,19.792 3.803,19.813 3.62,19.893C3.438,19.974 3.281,20.11 3.172,20.285C3.062,20.459 3.003,20.665 3.003,20.875C3.003,21.157 3.108,21.427 3.296,21.626C3.483,21.826 3.738,21.938 4.003,21.938ZM2.903,18.75H4.943L4.723,16.625H2.458L2.903,18.75ZM4.608,15.563L4.383,13.438H1.788L2.233,15.563H4.608Z"
+      android:fillColor="#ffffff"/>
+</vector>
