+++ presentation/src/main/res/drawable/ic_delete.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="40dp"
+    android:height="40dp"
+    android:viewportWidth="40"
+    android:viewportHeight="40">
+  <path
+      android:pathData="M28,13.611L26.389,12L20,18.389L13.611,12L12,13.611L18.389,20L12,26.389L13.611,28L20,21.611L26.389,28L28,26.389L21.611,20L28,13.611Z"
+      android:fillColor="#AAAAAA"/>
+</vector>
