+++ presentation/src/main/res/values/strings.xml
@@ -1,0 +1,14 @@
+<resources>
+    <string name="app_name">Shopping</string>
+
+    <string name="product_price">%,d원</string>
+
+    <string name="cart_title">장바구니</string>
+    <string name="delete_cart_product">삭제</string>
+    <string name="menu_cart_title">장바구니</string>
+
+    <string name="date_format">d MMM yyyy HH:mm:ss</string>
+
+    <string name="cart_added">장바구니에 추가되었습니다.</string>
+    <string name="cart_deleted">품목이 삭제되었습니다.</string>
+</resources>
