+++ settings.gradle.kts
@@ -16,3 +16,4 @@ rootProject.name = "android-di"
 include(":app")
 include(":data")
 include(":local")
+include(":presentation")
